day 1 of refactoring complete
can instantiate monsters and it all works a lot more seamlessly. i use custom class dict to send monster data resources and levels to the enemy party

that way the data is never harmed and we're always able to pull from it.

learned a lot more about signals on my first go around of this. about 2 weeks deep in godot